chore: more reorganization of notify.go
